fix: roles.service

+++ src/roles/roles.service.ts
@@ -1,6 +1,9 @@
 import {
   BadRequestException,
+  ConflictException,
+  ForbiddenException,
   Injectable,
+  Logger,
   NotFoundException,
 } from '@nestjs/common';
 import { Repository } from 'typeorm';
@@ -18,6 +21,7 @@ import { UpdateRoleDto } from './dto/update-role.dto';
  */
 @Injectable()
 export class RolesService {
+  private readonly logger = new Logger('RolesService');
   /**
    * Constructor initializing repository dependency
    * @param roleRepository Repository for performing database operations on roles
@@ -34,17 +38,23 @@ export class RolesService {
    * @throws BadRequestException if role with the same name already exists
    */
   async create(createRoleDto: CreateRoleDto) {
-    const role = await this.roleRepository.findOne({
-      where: { name: createRoleDto.name },
-    });
+    try {
+      const role = await this.roleRepository.findOne({
+        where: { name: createRoleDto.name },
+      });
 
-    if (role) {
-      throw new BadRequestException(
-        `role with name ${createRoleDto.name} exist`,
-      );
+      if (role) {
+        throw new BadRequestException(
+          `role with name ${createRoleDto.name} exist`,
+        );
+      }
+
+      return this.roleRepository.save(createRoleDto);
+    } catch (error) {
+      this.logger.error('Find roles error');
+
+      throw error;
     }
-
-    return this.roleRepository.save(createRoleDto);
   }
 
   /**
@@ -52,12 +62,18 @@ export class RolesService {
    * @returns Array of roles with selected fields (id and name)
    */
   findAll() {
-    return this.roleRepository.find({
-      select: {
-        id: true,
-        name: true,
-      },
-    });
+    try {
+      return this.roleRepository.find({
+        select: {
+          id: true,
+          name: true,
+        },
+      });
+    } catch (error) {
+      this.logger.error('Find roles error');
+
+      throw error;
+    }
   }
 
   /**
@@ -67,19 +83,29 @@ export class RolesService {
    * @throws NotFoundException if role with given ID does not exist
    */
   async findOne(id: number) {
-    const role = await this.roleRepository.findOne({
-      where: { id },
-      select: {
-        id: true,
-        name: true,
-      },
-    });
+    try {
+      if (typeof id !== 'number' || id <= 0) {
+        throw new BadRequestException('Invalid role ID');
+      }
 
-    if (!role) {
-      throw new NotFoundException(`role with id ${id} not found`);
+      const role = await this.roleRepository.findOne({
+        where: { id },
+        select: {
+          id: true,
+          name: true,
+        },
+      });
+
+      if (!role) {
+        throw new NotFoundException(`role with id ${id} not found`);
+      }
+
+      return role;
+    } catch (error) {
+      this.logger.error('Find role error');
+
+      throw error;
     }
-
-    return role;
   }
 
   /**
@@ -88,8 +114,37 @@ export class RolesService {
    * @param updateRoleDto DTO containing updated role data
    * @returns Update result
    */
-  update(id: number, updateRoleDto: UpdateRoleDto) {
-    return this.roleRepository.update(+id, updateRoleDto);
+  async update(id: number, updateRoleDto: UpdateRoleDto) {
+    try {
+      if (typeof id !== 'number' || id <= 0) {
+        throw new BadRequestException('Invalid role ID');
+      }
+
+      const existingRole = await this.roleRepository.findOne({
+        where: { id },
+        relations: ['users'],
+      });
+
+      if (!existingRole) {
+        throw new NotFoundException('Role not found');
+      }
+
+      if (updateRoleDto.name && updateRoleDto.name !== existingRole.name) {
+        const roleWithSameName = await this.roleRepository.findOne({
+          where: { name: updateRoleDto.name.toLowerCase() },
+        });
+
+        if (roleWithSameName) {
+          throw new ConflictException('A role with this name already exists');
+        }
+      }
+
+      return this.roleRepository.update(+id, updateRoleDto);
+    } catch (error) {
+      this.logger.error('Update role error');
+
+      throw error;
+    }
   }
 
   /**
@@ -97,7 +152,34 @@ export class RolesService {
    * @param id ID of the role to delete
    * @returns Deletion result
    */
-  remove(id: number) {
-    return this.roleRepository.delete(id);
+  async remove(id: number) {
+    try {
+      if (!Number.isInteger(id) || id <= 0) {
+        throw new BadRequestException('Invalid role ID');
+      }
+
+      const role = await this.roleRepository.findOne({
+        where: { id },
+        relations: ['users'], // Проверяем связанные пользователи
+      });
+
+      if (!role) {
+        throw new NotFoundException('Role not found');
+      }
+
+      if (role.users.length > 0) {
+        throw new ForbiddenException(
+          'The role cannot be deleted because it is associated with users'
+        );
+      }
+
+      this.logger.log(`Delete role ${id}`);
+
+      return this.roleRepository.delete(id);
+    } catch (error) {
+      this.logger.error('Delete role error');
+
+      throw error;
+    }
   }
 }
